requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
Pillow
social-auth-app-django
django-environ
requests
psycopg2-binary
django-debug-toolbar
social_auth_app_django
django-debug-toolbar-template-profiler
